Move resolvePath to its own file
