+++ client/src/extension.test.ts
@@ -1,5 +1,23 @@
-import { expect, test } from 'vitest';
+import { expect, test, describe, vi, beforeAll, afterAll } from 'vitest';
+import { resolvePath } from './utils';
 
-test('adds 1 + 2 to equal 3', () => {
-	expect(1 + 2).toBe(3);
+describe('resolvePath', () => {
+	beforeAll(() => {
+		// Mock the behavior of fs.lstatSync
+		vi.mock('fs', () => ({
+			lstatSync: (_path: string) => ({ isSymbolicLink: () => false }),
+		}));
+	});
+
+	afterAll(() => {
+		vi.restoreAllMocks();
+	});
+
+	test.concurrent.each([
+		['', ''],
+		['C:/out.txt', 'C:/out.txt'],
+	])('("%s") -> "%s"', (input, expected) => {
+		const result = resolvePath(input);
+		expect(result).toBe(expected);
+	});
 });
